backend/python/requirements.txt
# PBI Scheduler Dependencies
# Install with: pip install -r requirements-scheduler.txt

# =============================================================================
# Core Scheduling
# =============================================================================
APScheduler>=3.10.4
croniter>=2.0.1

# =============================================================================
# CLI Interface
# =============================================================================
click>=8.1.7
rich>=13.7.0

# =============================================================================
# Web UI
# =============================================================================
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Werkzeug>=3.0.0

# =============================================================================
# Authentication
# =============================================================================
Authlib>=1.3.0
bcrypt>=4.1.2

# =============================================================================
# Production Server
# =============================================================================
gunicorn>=21.2.0

# =============================================================================
# Alerting
# =============================================================================
slack-sdk>=3.26.0

# =============================================================================
# Configuration
# =============================================================================
PyYAML>=6.0.1
python-decouple>=3.8
python-dotenv>=1.0.0

# =============================================================================
# Database
# =============================================================================
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# =============================================================================
# HTTP/API Clients
# =============================================================================
requests>=2.31.0
urllib3>=2.0.0

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.1.0
tqdm>=4.66.0
python-dateutil>=2.8.2

# =============================================================================
# Caching
# =============================================================================
redis>=5.0.0

# =============================================================================
# Legacy Support (optional - for SLDB sync)
# =============================================================================
sshtunnel>=0.4.0
marshmallow>=3.20.0
PyJWT>=2.8.0
yfinance
